+++ lib/interfaces/ProcessDownloadFileParams.ts
@@ -2,4 +2,5 @@ export interface ProcessDownloadFileParams {
 	asyncDownload?: boolean;
 	pollInitialWaitTime?: number;
 	pollMaximumWaitTime?: number;
+	bundleDownloadTimeout?: number;
 }